Refactor code into packages and update test file
