+++ src/main/java/bernard/core/Ui.java
@@ -1,3 +1,5 @@
+package bernard.core;
+
 import java.util.Scanner;
 
 public class Ui {
