+++ src/main/java/bernard/exceptions/BernardException.java
@@ -1,3 +1,5 @@
+package bernard.exceptions;
+
 public class BernardException extends Exception{
     public BernardException(String message) {
         super(message);
