+++ src/main/java/bernard/tasks/Todo.java
@@ -1,3 +1,7 @@
+package bernard.tasks;
+
+import bernard.exceptions.BernardException;
+
 public class Todo extends Task {
     public Todo(String description) throws BernardException {
         super(description);
